Modify Test and NOLINT
